--- a/frappe/core/doctype/version/version.py
+++ b/frappe/core/doctype/version/version.py
@@ -2,6 +2,7 @@
 # License: MIT. See LICENSE
 
 import json
+from typing import Optional
 
 import frappe
 from frappe.model import no_value_fields, table_fields
@@ -9,7 +10,15 @@ from frappe.model.document import Document
 
 
 class Version(Document):
-	def set_diff(self, old, new):
+	def update_version_info(self, old: Optional[Document], new: Document) -> bool:
+		"""Update changed info and return true if change contains useful data."""
+		if not old:
+			# Check if doc has some information about creation source like data import
+			return self.for_insert(new)
+		else:
+			return self.set_diff(old, new)
+
+	def set_diff(self, old: Document, new: Document) -> bool:
 		"""Set the data property with the diff of the docs if present"""
 		diff = get_diff(old, new)
 		if diff:
@@ -20,8 +29,11 @@ class Version(Document):
 		else:
 			return False
 
-	def for_insert(self, doc):
+	def for_insert(self, doc: Document) -> bool:
 		updater_reference = doc.flags.updater_reference
+		if not updater_reference:
+			return False
+
 		data = {
 			"creation": doc.creation,
 			"updater_reference": updater_reference,
@@ -29,7 +41,8 @@ class Version(Document):
 		}
 		self.ref_doctype = doc.doctype
 		self.docname = doc.name
-		self.data = frappe.as_json(data)
+		self.data = frappe.as_json(data, indent=None, separators=(",", ":"))
+		return True
 
 	def get_data(self):
 		return json.loads(self.data)