Each new doc inserts a version, this contains nothing but creator and
creation time.. which is already immutable information on the original
document.

This was added for cases like data import to track from where document
got created, ref: https://github.com/frappe/frappe/commit/b7dfe7969dd6b2851844b1e9c090cbb6447748de

Fix: only add a version on creation IF creation info is present on flags

---

Testing inserts on simple doctypes
```python
def insertmany():
    for _ in range(10):
        t = frappe.new_doc("ToDo")
        t.description = "a"
        t.save()
```

Before:
`71 ms ± 1.44 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)`

After:

`49 ms ± 883 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)` ~30% faster and consumes less DB storage 🚀